+++ fenbux/univariate/continuous/_wald.py
@@ -147,7 +147,7 @@ def _quantile_impl(dist: Wald, x: PyTreeVar):
 
 @_rand_impl.dispatch
 def _rand(
-    d: Wald, key: KeyArray, shape: Shape = (), dtype: DTypeLikeFloat = jnp.float_
+    d: Wald, key: KeyArray, shape: Shape = (), dtype: DTypeLikeFloat = float
 ):
     d = d.broadcast_params()
     _key_tree = split_tree(key, d.mu)
